GEOS-3830) UpdateSequence is not getting updated
git-svn-id: http://svn.codehaus.org/geoserver/trunk/src@14062 ef1d6e69-97e5-0310-af46-8a06194da32a

Former-commit-id: 78a537aea8157b06e55e0bb4d10794be0089b781 [formerly c64b27a0fb9d69094a5a06debf9715e2e29d50db]
Former-commit-id: fbbe5e4e08f103ccedb5d4a4dd36de9491420de8

+++ web/core/src/main/java/org/geoserver/web/admin/StatusPage.java
@@ -149,7 +149,7 @@ public class StatusPage extends ServerAdminPage {
         values.put(KEY_JAI_TILE_THREAD_PRIORITY, Integer.toString(jai.getTileScheduler()
                 .getPriority()));
 
-        values.put(KEY_UPDATE_SEQUENCE, Integer.toString(geoServerInfo.getUpdateSequence()));
+        values.put(KEY_UPDATE_SEQUENCE, Long.toString(geoServerInfo.getUpdateSequence()));
     }
 
     boolean isNativeJAIAvailable() {
